+++ Example/SJSegmentedScrollViewDemo/ViewControllers/ViewController.swift
@@ -45,11 +45,13 @@ class ViewController: SJSegmentedViewController {
 			                           fourthViewController]
 			headerViewHeight = 200
 			selectedSegmentViewHeight = 5.0
+			headerViewOffsetHeight = 31.0
 			segmentTitleColor = .gray
 			selectedSegmentViewColor = .red
 			segmentShadow = SJShadow.light()
 			showsHorizontalScrollIndicator = false
 			showsVerticalScrollIndicator = false
+			segmentBounces = false
 			delegate = self
 		}