+++ CHANGELOG.md
@@ -42,3 +42,9 @@ All notable changes to this project will be documented in this file.
 
 #### `1.3.6`
 * Resolved issue #54. How to disable Scroll indicator.
+
+#### `1.3.8`
+* Migrated to Swift 4.
+
+#### `1.3.9`
+* Added 'segmentBounces' property to set content view bounces.